convert workload virtual address to physical address

+++ criu/include/vptrans.h
@@ -5,9 +5,9 @@
 #include "common/list.h"
 
 struct vptrans_pin {
-	unsigned long vaddr;
-	u64 off;
-	u64 nr_page;
+	uintptr_t vaddr;
+	unsigned int off;
+	int nr_page;
 };
 
 #endif
